requirements.txt
numpy==1.23.4
pwntools==4.8.0
